+++ WareHouseRT/src/main/java/WareHouseRT/WareHouseRT/repository/DesktopRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import WareHouseRT.WareHouseRT.beans.Desktop;
 
 
-public interface DesktopRepository extends MongoRepository<Desktop, Integer>{
+public interface DesktopRepository extends MongoRepository<Desktop, Long>{
 
 }
