+++ WareHouseRT/src/main/java/WareHouseRT/WareHouseRT/repository/GpuRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import WareHouseRT.WareHouseRT.beans.GPU;
 
-public interface GpuRepository extends MongoRepository<GPU, Integer>{
+public interface GpuRepository extends MongoRepository<GPU, Long>{
 
 }
