+++ WareHouseRT/src/main/java/WareHouseRT/WareHouseRT/repository/HddRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import WareHouseRT.WareHouseRT.beans.HDD;
 
-public interface HddRepository extends MongoRepository<HDD, Integer>{
+public interface HddRepository extends MongoRepository<HDD, Long>{
 
 }
