+++ WareHouseRT/src/main/java/WareHouseRT/WareHouseRT/repository/LaptopRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import WareHouseRT.WareHouseRT.beans.Laptop;
 
-public interface LaptopRepository extends MongoRepository<Laptop, Integer>{
+public interface LaptopRepository extends MongoRepository<Laptop, Long>{
 
 }
